% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance fire spread}
  \item{DAMAGE}{Damage done by fire}
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set about wildfires and damage
}
\details{
The damage of fire is measured.
}
\keyword{datasets}
