% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of times we resample (with replacement)}

\item{x}{original sample data}

\item{fun}{If we want to approx fun(population), fun is the function "fun"}

\item{alpha}{value of alpha for conf int}

\item{cx}{Text size for histogram}

\item{...}{Other graphical parameters}
}
\value{
prints Histogram of bootstrap resample distribution with estimate of fun(population) with a 100(1-a)% confidence interval
}
\description{
myboot2
}
\examples{
myboot2(iter=10000,c(1,1,1,1,3,3,3,3,3,3,4,5,1,1,1,6,2,2,2,2),fun = median, alpha = 0.20)
}
