% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{mycltb}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{number of terms in each sample}

\item{iter}{number of samples (should be large to mimic infinity)}

\item{p}{probability of success in each binomial trial}

\item{...}{extra}
}
\value{
Makes a sample from the binomial distribution given, and outputs the sample's mean and a graph of its mean.
}
\description{
mycltb
}
\examples{
mycltb(n=10,iter=10000,p=0.5)
}
