% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{My first function}
\usage{
myfun(x)
}
\arguments{
\item{x}{A quantitative vector #What is the input}
}
\value{
A quantitative vector, squares param #what does it return?
}
\description{
My first function
}
\examples{
myfun(x = 1:10)
}
