% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{myhyper}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{How many samples}

\item{N}{Total size of population}

\item{r}{number of possible successes in population}

\item{n}{Total size of each sample taken from population}
}
\value{
creates barplot of a hypergeometric trial
}
\description{
myhyper
}
\examples{
myhyper(iter=100,N=20,r=12,n=5)
}
