% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{upper limit of pnorm}
}
\value{
Plots normal curve with mean = mu, sd = sigma, shades area from -inf to a, and prints the area.
}
\description{
myncurve
}
\examples{
myncurve(0,1,0.5)
}
