% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.r
\name{pareto}
\alias{pareto}
\title{Pareto Barplot Function}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{...}{}
}
\value{
returns Pareto Barplot of x with title mn
}
\description{
Pareto Barplot Function
}
