% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.r
\name{pareto}
\alias{pareto}
\title{Pareto Barplot Function}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{data}

\item{mn}{main name}

\item{...}{other stuff}
}
\value{
returns Pareto Barplot of x with title mn
}
\description{
Pareto Barplot Function
}
\examples{
pareto(c(1,2,3,4),mn="Pareto Barplot")
}
